+++ src/lib/warn/warn.ts
@@ -1,0 +1,9 @@
+let enable = true;
+
+export const disableWarn = () => {
+  enable = false;
+};
+
+export default (...args) => {
+  enable && console.warn(...args);
+};