+++ test/e2e/helpers/getValidation.js
@@ -9,13 +9,13 @@ export default ({
     options.validateOnEvents = validateOnEvents;
   }
 
-  const { createForm, validate, clearErrors, validateValueStore, createEntry, createEntries } = createValidation({
+  const { createForm, validate, clearErrors, validateValueStore, createEntry, createEntries, getValues } = createValidation({
     presence: 'required', // set required for e2e testing
     ...options
   });
 
   return {
     entries: createEntries(entries),
-    createForm, validate, clearErrors, validateValueStore, createEntry
+    createForm, validate, clearErrors, validateValueStore, createEntry, getValues
   }
 }